backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.6.0
djongo==1.3.6
python-decouple==3.8
requests==2.31.0
bcrypt==4.1.1
PyJWT==2.8.0